Revamp UI and improve security
