+++ README.md
@@ -1,75 +1,25 @@
-# Nuxt 3 Minimal Starter
+# BrowserJam
 
-Look at the [Nuxt 3 documentation](https://nuxt.com/docs/getting-started/introduction) to learn more.
+BrowserJam is a Nuxt 3 application that allows you to play **Animal Jam Classic** in the browser using the Ruffle Flash emulator. The interface is intentionally minimal – log in with your Animal Jam credentials and start playing.
+
+This project currently supports only the Play feature. Other UI elements from previous versions have been removed.
 
 ## Setup
 
-Make sure to install the dependencies:
+Install dependencies and start the development server:
 
 ```bash
-# npm
 npm install
-
-# pnpm
-pnpm install
-
-# yarn
-yarn install
-
-# bun
-bun install
-```
-
-## Development Server
-
-Start the development server on `http://localhost:3000`:
-
-```bash
-# npm
 npm run dev
-
-# pnpm
-pnpm run dev
-
-# yarn
-yarn dev
-
-# bun
-bun run dev
 ```
 
-## Production
+### Production build
 
-Build the application for production:
+To build and preview the production version:
 
 ```bash
-# npm
 npm run build
-
-# pnpm
-pnpm run build
-
-# yarn
-yarn build
-
-# bun
-bun run build
-```
-
-Locally preview production build:
-
-```bash
-# npm
 npm run preview
-
-# pnpm
-pnpm run preview
-
-# yarn
-yarn preview
-
-# bun
-bun run preview
 ```
 
-Check out the [deployment documentation](https://nuxt.com/docs/getting-started/deployment) for more information.
+The app is deployed on [Railway](https://animaljam.up.railway.app/).
