+++ src/server/api/image.ts
@@ -4,9 +4,20 @@ import { Buffer } from 'buffer';
 export default defineEventHandler(async (event) => {
   const query = getQuery(event);
   const url = query.url as string;
-  
+
   if (!url) {
     return createError({ statusCode: 400, message: 'Missing url parameter' });
+  }
+
+  let parsedUrl: URL
+  try {
+    parsedUrl = new URL(url)
+  } catch {
+    return createError({ statusCode: 400, message: 'Invalid url parameter' })
+  }
+
+  if (parsedUrl.hostname !== 'ajcontent.akamaized.net') {
+    return createError({ statusCode: 400, message: 'URL not allowed' })
   }
   
   try {
