+++ tailwind.config.js
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: [],
+  content: ['./src/**/*.{vue,js,ts}'],
   theme: {},
   plugins: [],
 }
